Backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.0
sqlalchemy>=2.0.36
psycopg[binary]
passlib
bcrypt
pyjwt
python-multipart
python-dotenv==1.0.0
websockets==12.0
alembic==1.13.1
google-generativeai>=0.8.0